fix(lib): types.h u32/s32 use long not int after chantier A1
Companion fix for chantier A1 (commit 6eb62e1). The CI's full test
suite (which runs Compiler Tests phase, skipped by my local --quick)
caught two cascade issues from the int/long size change:

1. **`lib/include/snes/types.h:59,66`** — `u32` and `s32` were
   typedef'd to `unsigned int` / `signed int` with the documented
   rationale "Uses int because cproc treats int as 32-bit on 65816".
   Pre-A1 that worked: `int=4`, so `u32=4`. Post-A1: `int=2`, so
   `u32=2` — silently halving the storage of every `u32` variable
   across the SDK and user code.

   Fix: `u32 = unsigned long` (`long=4` post-A1), `s32 = signed long`.
   Comment block updated to reflect post-A1 reality.

2. **`tools/opensnes-emu` test fixture** — same issue in
   `test/fixtures/compiler/test_u32_arithmetic.c`. Submodule bumped
   to e8a7e9d to pick up the matching fix.

Both fixes are mechanical consequences of the A1 type-size change
that I missed during local validation because `--quick` mode skips
the Compiler Tests phase. The CI's full suite caught them within
3 minutes of pushing 6eb62e1.

Validation: full suite with `--allow-known-bugs` (matching CI config)
returns 404/404 under the combined parent + submodule state. No
remaining cascade items expected — the only place in the lib that
relied on `int=4` was the `u32`/`s32` typedef, and the only test
fixture that hardcoded the old typedef locally was
test_u32_arithmetic.c.

Status note: A1 chantier is now FULLY shipped (was PARTIAL after
6eb62e1; complete after this commit). The 1 remaining "known bug"
in the suite (`nonleaf_frameless`) predates A1 and is the cosmetic
`leaf_opt=1` marker drift documented in KNOWN_LIMITATIONS.md.

+++ lib/include/snes/types.h
@@ -54,16 +54,19 @@ typedef unsigned short u16;
 /**
  * @brief 32-bit signed integer (-2147483648 to 2147483647)
  *
- * @note Uses `int` because cproc treats `int` as 32-bit on 65816.
+ * Uses `long` (4 bytes on cc65816 since chantier A1, 2026-05-08). Pre-A1
+ * this was `int` because cproc treated `int` as 32-bit; with A1 aligning
+ * `int` to 2 bytes (the native w65816 word size), the canonical 32-bit
+ * type moves to `long`.
  */
-typedef signed int s32;
+typedef signed long s32;
 
 /**
  * @brief 32-bit unsigned integer (0 to 4294967295)
  *
- * @note Uses `int` because cproc treats `int` as 32-bit on 65816.
+ * Uses `long` (4 bytes on cc65816 since chantier A1, 2026-05-08).
  */
-typedef unsigned int u32;
+typedef unsigned long u32;
 
 /** @} */ /* end of types group */
